move argonaut tests
