+++ lib/modules/argonaut/test/argonaut_allergy_intolerance_sequence_test.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
-class ArgonautPatientSearchSequenceTest < MiniTest::Test
+require_relative '../../../../test/test_helper'
+class ArgonautAllergyIntoleranceSequenceTest < MiniTest::Test
   def setup
     @instance = get_test_instance
     client = get_client(@instance)
 
-    @fixture = 'patient' # put fixture file name here
-    @sequence = Inferno::Sequence::ArgonautPatientSequence.new(@instance, client) # put sequence here
-    @resource_type = 'Patient'
+    @fixture = 'allergy_intolerance'
+    @sequence = Inferno::Sequence::ArgonautAllergyIntoleranceSequence.new(@instance, client)
+    @resource_type = 'AllergyIntolerance'
 
     @resource = FHIR::DSTU2.from_contents(load_fixture(@fixture.to_sym))
     assert_empty @resource.validate, "Setup failure: Resource fixture #{@fixture}.json not a valid #{@resource_type}."
@@ -19,11 +19,10 @@ class ArgonautPatientSearchSequenceTest < MiniTest::Test
       entry.resource.meta.versionId = '1'
     end
 
-    @patient_id = @resource.id
+    @patient_id = @resource.patient.reference
     @patient_id = @patient_id.split('/')[-1] if @patient_id.include?('/')
 
     @patient_resource = FHIR::DSTU2::Patient.new(id: @patient_id)
-    @practitioner_resource = FHIR::DSTU2::Practitioner.new(id: 432)
 
     # Assume we already have a patient
     @instance.resource_references << Inferno::Models::ResourceReference.new(
@@ -38,8 +37,7 @@ class ArgonautPatientSearchSequenceTest < MiniTest::Test
     @request_headers = { 'Accept' => 'application/json+fhir',
                          'Accept-Charset' => 'utf-8',
                          'User-Agent' => 'Ruby FHIR Client',
-                         'Authorization' => "Bearer #{@instance.token}",
-                         'Host' => 'www.example.com' }
+                         'Authorization' => "Bearer #{@instance.token}" }
 
     @extended_request_headers = { 'Accept' => 'application/json+fhir',
                                   'Accept-Charset' => 'utf-8',
@@ -55,7 +53,6 @@ class ArgonautPatientSearchSequenceTest < MiniTest::Test
     # Return 401 if no Authorization Header
     uri_template = Addressable::Template.new "http://www.example.com/#{@resource_type}{?patient,target,start,end,userid,agent}"
     stub_request(:get, uri_template).to_return(status: 401)
-    stub_request(:get, "http://www.example.com/#{@resource_type}/#{@resource.id}").to_return(status: 401)
 
     # Search Resources
     stub_request(:get, uri_template)
@@ -83,6 +80,15 @@ class ArgonautPatientSearchSequenceTest < MiniTest::Test
       .with(headers: @extended_request_headers)
       .to_return(status: 200,
                  body: @resource.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+
+    # Stub Patient for Reference Resolution Tests
+    stub_request(:get, %r{example.com/Patient/})
+      .with(headers: {
+              'Authorization' => "Bearer #{@instance.token}"
+            })
+      .to_return(status: 200,
+                 body: @patient_resource.to_json,
                  headers: { content_type: 'application/json+fhir; charset=UTF-8' })
   end
