+++ lib/modules/argonaut/test/argonaut_careteam_sequence_test.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
-class ArgonautImmunizationTest < MiniTest::Test
+require_relative '../../../../test/test_helper'
+class CareTeamSequenceTest < MiniTest::Test
   def setup
     @instance = get_test_instance
     client = get_client(@instance)
 
-    @fixture = 'immunization' # put fixture file name here
-    @sequence = Inferno::Sequence::ArgonautImmunizationSequence.new(@instance, client) # put sequence here
-    @resource_type = 'Immunization'
+    @fixture = 'care_team' # put fixture file name here
+    @sequence = Inferno::Sequence::ArgonautCareTeamSequence.new(@instance, client) # put sequence here
+    @resource_type = 'CarePlan'
 
     @resource = FHIR::DSTU2.from_contents(load_fixture(@fixture.to_sym))
     assert_empty @resource.validate, "Setup failure: Resource fixture #{@fixture}.json not a valid #{@resource_type}."
@@ -19,11 +19,10 @@ class ArgonautImmunizationTest < MiniTest::Test
       entry.resource.meta.versionId = '1'
     end
 
-    @patient_id = @resource.patient.reference
+    @patient_id = @resource.subject.reference
     @patient_id = @patient_id.split('/')[-1] if @patient_id.include?('/')
 
     @patient_resource = FHIR::DSTU2::Patient.new(id: @patient_id)
-    @practitioner_resource = FHIR::DSTU2::Practitioner.new(id: 432)
 
     # Assume we already have a patient
     @instance.resource_references << Inferno::Models::ResourceReference.new(
@@ -52,7 +51,7 @@ class ArgonautImmunizationTest < MiniTest::Test
 
   def full_sequence_stubs
     # Return 401 if no Authorization Header
-    uri_template = Addressable::Template.new "http://www.example.com/#{@resource_type}{?patient,target,start,end,userid,agent}"
+    uri_template = Addressable::Template.new "http://www.example.com/#{@resource_type}{?patient,category}"
     stub_request(:get, uri_template).to_return(status: 401)
 
     # Search Resources
@@ -91,6 +90,42 @@ class ArgonautImmunizationTest < MiniTest::Test
       .to_return(status: 200,
                  body: @patient_resource.to_json,
                  headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+    stub_request(:get, %r{example.com/Practitioner/1})
+      .with(headers: @extended_request_headers)
+      .to_return(status: 200,
+                 body: @patient_resource.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+  end
+
+  def wrong_resource_stubs
+    # Return the wrong profile
+    uri_template = Addressable::Template.new "http://www.example.com/#{@resource_type}{?patient,category}"
+    wrong_resource = FHIR::DSTU2.from_contents(load_fixture('care_plan'))
+
+    wrong_resource_bundle = wrap_resources_in_bundle(wrong_resource)
+
+    wrong_resource_bundle.entry.each do |entry|
+      entry.resource.meta = FHIR::DSTU2::Meta.new unless entry.resource.meta
+      entry.resource.meta.versionId = '1'
+    end
+
+    stub_request(:get, uri_template)
+      .with(headers: @request_headers)
+      .to_return(
+        status: 200, body: wrong_resource_bundle.to_json, headers: @response_headers
+      )
+
+    stub_request(:get, "http://www.example.com/#{@resource_type}/#{@resource.id}")
+      .with(headers: @request_headers)
+      .to_return(status: 200,
+                 body: wrong_resource.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+
+    stub_request(:get, %r{example.com/Patient/})
+      .with(headers: { 'Authorization' => "Bearer #{@instance.token}" })
+      .to_return(status: 200,
+                 body: @patient_resource.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
   end
 
   def test_all_pass
@@ -98,11 +133,26 @@ class ArgonautImmunizationTest < MiniTest::Test
 
     sequence_result = @sequence.start
 
-    # Currently will have a warning about not finding the valueset
-    # TODO: FIX THIS WARNING
     failures = sequence_result.failures
     assert failures.empty?, "All tests should pass.  First error: #{!failures.empty? && failures.first.message}"
+    sequence_result.test_results.each do |test_result|
+      assert test_result.pass?, "#{test_result.name} - #{test_result.result}"
+    end
     assert sequence_result.pass?, "The sequence should be marked as pass. #{sequence_result.result}"
-    # assert sequence_result.test_results.all? { |r| r.test_warnings.empty? }, 'There should not be any warnings.'
+    assert sequence_result.test_results.all? { |r| r.test_warnings.empty? }, 'There should not be any warnings.'
+  end
+
+  # Return a careplan profiled resource even though we are asking for a careteam profile
+  # This should have at least one failure
+  def test_fail_on_wrong_profile
+    wrong_resource_stubs
+
+    sequence_result = @sequence.start
+
+    failures = sequence_result.test_results.select(&:fail?)
+
+    assert failures.present?, 'Expecting at least one failure because we are returning a resource stating conformance to the wrong profile'
+    assert sequence_result.result == 'fail', "The sequence should be marked as fail. #{sequence_result.result}"
+    assert sequence_result.test_results.all? { |r| r.test_warnings.empty? }, 'There should not be any warnings.'
   end
 end
