+++ lib/modules/argonaut/test/argonaut_medication_order_sequence_test.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
-class DiagnosticReportTest < MiniTest::Test
+require_relative '../../../../test/test_helper'
+class ArgonautMedicationOrderSequenceTest < MiniTest::Test
   def setup
     @instance = get_test_instance
     client = get_client(@instance)
 
-    @fixture = 'diagnostic_report' # put fixture file name here
-    @sequence = Inferno::Sequence::ArgonautDiagnosticReportSequence.new(@instance, client) # put sequence here
-    @resource_type = 'DiagnosticReport'
+    @fixture = 'medication_order' # put fixture file name here
+    @sequence = Inferno::Sequence::ArgonautMedicationOrderSequence.new(@instance, client) # put sequence here
+    @resource_type = 'MedicationOrder'
 
     @resource = FHIR::DSTU2.from_contents(load_fixture(@fixture.to_sym))
     assert_empty @resource.validate, "Setup failure: Resource fixture #{@fixture}.json not a valid #{@resource_type}."
@@ -19,13 +19,13 @@ class DiagnosticReportTest < MiniTest::Test
       entry.resource.meta.versionId = '1'
     end
 
-    @patient_id = @resource.subject.reference
+    @patient_id = @resource.patient.reference
     @patient_id = @patient_id.split('/')[-1] if @patient_id.include?('/')
 
     @patient_resource = FHIR::DSTU2::Patient.new(id: @patient_id)
-    @encounter_resource = FHIR::DSTU2::Encounter.new(id: 432)
     @practitioner_resource = FHIR::DSTU2::Practitioner.new(id: 432)
 
+    @medication_reference = load_json_fixture(:medication_reference)
     # Assume we already have a patient
     @instance.resource_references << Inferno::Models::ResourceReference.new(
       resource_type: 'Patient',
@@ -63,15 +63,15 @@ class DiagnosticReportTest < MiniTest::Test
         status: 200, body: @resource_bundle.to_json, headers: @response_headers
       )
 
-    stub_request(:get, %r{http\://www\.example\.com/DiagnosticReport\?.*date\=.*})
-      .with(headers: @extended_request_headers)
-      .to_return(
-        status: 200, body: @resource_bundle.to_json, headers: @response_headers
-      )
-
     # Read Resources
     stub_request(:get, "http://www.example.com/#{@resource_type}/#{@resource.id}")
       .with(headers: @request_headers)
+      .to_return(status: 200,
+                 body: @resource.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+
+    stub_request(:get, "http://www.example.com/#{@resource_type}/#{@resource.id}")
+      .with(headers: @extended_request_headers)
       .to_return(status: 200,
                  body: @resource.to_json,
                  headers: { content_type: 'application/json+fhir; charset=UTF-8' })
@@ -98,20 +98,21 @@ class DiagnosticReportTest < MiniTest::Test
       .to_return(status: 200,
                  body: @patient_resource.to_json,
                  headers: { content_type: 'application/json+fhir; charset=UTF-8' })
-
+    # Return Medication from a reference
+    stub_request(:get, %r{example.com/Medication/})
+      .with(headers: {
+              'Authorization' => "Bearer #{@instance.token}"
+            })
+      .to_return(status: 200,
+                 body: @medication_reference.to_json,
+                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
+    # Stub Practitioner for Reference Resolution Tests
     stub_request(:get, %r{example.com/Practitioner/})
       .with(headers: {
               'Authorization' => "Bearer #{@instance.token}"
             })
       .to_return(status: 200,
-                 body: @patient_resource.to_json,
-                 headers: { content_type: 'application/json+fhir; charset=UTF-8' })
-    stub_request(:get, %r{example.com/Encounter/})
-      .with(headers: {
-              'Authorization' => "Bearer #{@instance.token}"
-            })
-      .to_return(status: 200,
-                 body: @encounter_resource.to_json,
+                 body: @practitioner_resource.to_json,
                  headers: { content_type: 'application/json+fhir; charset=UTF-8' })
   end
