+++ lib/modules/argonaut/test/fixtures/allergy_intolerance.json
@@ -1,0 +1,46 @@
+ {
+    "resourceType": "AllergyIntolerance",
+    "id": "SMART-AllergyIntolerance-28",
+    "meta": {
+    "versionId": "3",
+    "lastUpdated": "2018-01-31T22:43:25.000+00:00"
+    },
+    "text": {
+    "status": "generated",
+    "div": "<div xmlns=\"http://www.w3.org/1999/xhtml\">Sensitivity to sulfonamide antibacterial</div>"
+    },
+    "recordedDate": "2000",
+    "patient": {
+    "reference": "Patient/SMART-1288992"
+    },
+    "substance": {
+    "coding": [
+        {
+        "system": "http://rxnav.nlm.nih.gov/REST/Ndfrt",
+        "code": "N0000175503",
+        "display": "sulfonamide antibacterial"
+        }
+    ],
+    "text": "sulfonamide antibacterial"
+    },
+    "status": "confirmed",
+    "criticality": "CRITL",
+    "category": "medication",
+    "reaction": [
+        {
+            "manifestation": [
+            {
+                "coding": [
+                {
+                    "system": "http://snomed.info/sct",
+                    "code": "271807003",
+                    "display": "skin rash"
+                }
+                ],
+                "text": "skin rash"
+            }
+            ],
+            "severity": "mild"
+        }
+    ]
+}
