+++ lib/modules/argonaut/test/fixtures/care_plan.json
@@ -1,0 +1,32 @@
+{
+  "id": "85e027e0-103b-0136-6943-60f81dc2d82a",
+  "meta": {
+    "profile": [
+      "http://fhir.org/guides/argonaut/StructureDefinition/argo-careplan"
+    ]
+  },
+  "text": {
+    "status": "additional",
+    "div": "This example is modified from the Official Argonaut Implementation Guide at http://www.fhir.org/guides/argonaut/r2/CarePlan-colonoscopy.html"
+  },
+  "subject": {
+    "reference": "Patient/7ec866ca-4130-4a40-82a9-b67f2626a5df",
+    "display": "Mrs. Rhea34 Ernser583"
+  },
+  "status": "active",
+  "period": {
+    "start": "2018-03-05",
+    "end": "2018-03-05"
+  },
+  "category": [
+    {
+      "coding": [
+        {
+          "system": "http://argonaut.hl7.org",
+          "code": "assess-plan"
+        }
+      ]
+    }
+  ],
+  "resourceType": "CarePlan"
+}
