+++ lib/modules/argonaut/test/fixtures/device.json
@@ -1,0 +1,28 @@
+{
+  "id": "8eb18b30-103b-0136-6943-60f81dc2d82a",
+  "meta": {
+    "profile": [
+      "http://fhir.org/guides/argonaut/StructureDefinition/argo-device"
+    ]
+  },
+  "text": {
+    "status": "generated",
+    "div": "This example is modified from the Official Argonaut Implementation Guide at http://www.fhir.org/guides/argonaut/r2/Device-udi-1.html"
+  },
+  "type": {
+    "coding": [
+      {
+        "system": "http://snomed.info/sct",
+        "code": "468063009",
+        "display": "Coated femoral stem prosthesis, modular"
+      }
+    ]
+  },
+  "status": "available",
+  "udi": "{01}09504000059118{17}141120{10}7654321D{21}10987654d321",
+  "patient": {
+    "reference": "Patient/7ec866ca-4130-4a40-82a9-b67f2626a5df",
+    "display": "Mrs. Rhea34 Ernser583"
+  },
+  "resourceType": "Device"
+}
