+++ lib/modules/argonaut/test/fixtures/goal.json
@@ -1,0 +1,20 @@
+{
+    "id": "2ec42e20-1046-0136-6944-60f81dc2d82a",
+    "meta": {
+    "profile": [
+        "http://fhir.org/guides/argonaut/StructureDefinition/argo-goal"
+    ]
+    },
+    "text": {
+    "status": "generated",
+    "div": "This example is modified from the Official Argonaut Implementation Guide at http://www.fhir.org/guides/argonaut/r2/Goal-goal-1.html"
+    },
+    "subject": {
+    "reference": "Patient/7ec866ca-4130-4a40-82a9-b67f2626a5df",
+    "display": "Mrs. Rhea34 Ernser583"
+    },
+    "description": "Patient is targeting a pulse oximetry of 92% and a weight of 195 lbs",
+    "status": "in-progress",
+    "statusDate": "2018-03-05",
+    "resourceType": "Goal"
+}
