+++ lib/modules/argonaut/test/fixtures/immunization.json
@@ -1,0 +1,72 @@
+{
+    "resourceType": "Immunization",
+    "id": "example",
+    "text": {
+      "status": "generated",
+      "div": "<div><p><b>Generated Narrative with Details</b></p><p><b>id</b>: example</p><p><b>identifier</b>: urn:oid:1.3.6.1.4.1.21367.2005.3.7.1234</p><p><b>status</b>: completed</p><p><b>date</b>: 10/01/2013</p><p><b>vaccineCode</b>: Fluvax (Influenza) <span>(Details : {urn:oid:1.2.36.1.2001.1005.17 code 'FLUVAX' = '??)</span></p><p><b>patient</b>: <a>Patient/example</a></p><p><b>wasNotGiven</b>: false</p><p><b>reported</b>: false</p><p><b>performer</b>: <a>Practitioner/example</a></p><p><b>requester</b>: <a>Practitioner/example</a></p><p><b>encounter</b>: <a>Encounter/example</a></p><p><b>manufacturer</b>: <a>Organization/hl7</a></p><p><b>location</b>: <a>Location/1</a></p><p><b>lotNumber</b>: AAJN11K</p><p><b>expirationDate</b>: 15/02/2015</p><p><b>site</b>: left arm <span>(Details : {http://hl7.org/fhir/v3/ActSite code 'LA' = 'left arm', given as 'left arm'})</span></p><p><b>route</b>: Injection, intramuscular <span>(Details : {http://hl7.org/fhir/v3/RouteOfAdministration code 'IM' = 'Injection, intramuscular', given as 'Injection, intramuscular'})</span></p><p><b>doseQuantity</b>: 5 mg<span> (Details: http://unitsofmeasure.org code mg = '??')</span></p><p><b>note</b>: Notes on adminstration of vaccine</p><h3>Explanations</h3><table><tr><td>-</td><td><b>Reason</b></td></tr><tr><td>*</td><td>429060002 <span>(Details : {SNOMED CT code '429060002' = '429060002)</span></td></tr></table><h3>Reactions</h3><table><tr><td>-</td><td><b>Date</b></td><td><b>Detail</b></td><td><b>Reported</b></td></tr><tr><td>*</td><td>10/01/2013</td><td><a>Observation/example</a></td><td>true</td></tr></table><h3>VaccinationProtocols</h3><table><tr><td>-</td><td><b>DoseSequence</b></td><td><b>Description</b></td><td><b>Authority</b></td><td><b>Series</b></td><td><b>SeriesDoses</b></td><td><b>TargetDisease</b></td><td><b>DoseStatus</b></td><td><b>DoseStatusReason</b></td></tr><tr><td>*</td><td>1</td><td>Vaccination Protocol Sequence 1</td><td><a>Organization/hl7</a></td><td>Vaccination Series 1</td><td>2</td><td>1857005 <span>(Details : {SNOMED CT code '1857005' = '1857005)</span></td><td>Counts <span>(Details : {http://hl7.org/fhir/vaccination-protocol-dose-status code 'count' = 'Counts', given as 'Counts'})</span></td><td>Cold chain break <span>(Details : {http://hl7.org/fhir/vaccination-protocol-dose-status-reason code 'coldchbrk' = 'Cold chain break', given as 'Cold chain break'})</span></td></tr></table></div>"
+    },
+    "identifier": [
+      {
+        "system": "urn:ietf:rfc:3986",
+        "value": "urn:oid:1.3.6.1.4.1.21367.2005.3.7.1234"
+      }
+    ],
+    "status": "completed",
+    "date": "2013-01-10",
+    "vaccineCode": {
+      "coding": [
+        {
+          "system": "urn:oid:1.2.36.1.2001.1005.17",
+          "code": "FLUVAX"
+        }
+      ],
+      "text": "Fluvax (Influenza)"
+    },
+    "patient": {
+      "reference": "Patient/example"
+    },
+    "wasNotGiven": false,
+    "reported": false,
+    "lotNumber": "AAJN11K",
+    "expirationDate": "2015-02-15",
+    "site": {
+      "coding": [
+        {
+          "system": "http://hl7.org/fhir/v3/ActSite",
+          "code": "LA",
+          "display": "left arm"
+        }
+      ]
+    },
+    "route": {
+      "coding": [
+        {
+          "system": "http://hl7.org/fhir/v3/RouteOfAdministration",
+          "code": "IM",
+          "display": "Injection, intramuscular"
+        }
+      ]
+    },
+    "doseQuantity": {
+      "value": 5,
+      "system": "http://unitsofmeasure.org",
+      "code": "mg"
+    },
+    "note": [
+      {
+        "text": "Notes on adminstration of vaccine"
+      }
+    ],
+    "explanation": {
+      "reason": [
+        {
+          "coding": [
+            {
+              "system": "http://snomed.info/sct",
+              "code": "429060002"
+            }
+          ]
+        }
+      ]
+    }
+  }
