+++ lib/modules/argonaut/test/fixtures/medication_reference.json
@@ -1,0 +1,17 @@
+{
+  "resourceType": "Medication",
+  "id": "Medication15",
+  "text": {
+    "status": "generated",
+    "div": "<div xmlns=\"http://www.w3.org/1999/xhtml\"><p><b>Generated Narrative with Details</b></p><p><b>id</b>: medexample015</p><p><b>contained</b>: , </p><p><b>code</b>: Capecitabine 500mg oral tablet (Xeloda) <span>(Details : {RxNorm code '213293' = '213293', given as 'Capecitabine 500mg oral tablet (Xeloda)'})</span></p><p><b>isBrand</b>: true</p><p><b>manufacturer</b>: id: org2; name: Gene Inc</p><p><b>form</b>: Tablet dose form (qualifier value) <span>(Details : {SNOMED CT code '385055001' = 'Tablet', given as 'Tablet dose form (qualifier value)'})</span></p><h3>Ingredients</h3><table><tr><td>-</td><td><b>Item[x]</b></td><td><b>Amount</b></td></tr><tr><td>*</td><td>id: sub04; Capecitabine (substance) <span>(Details : {SNOMED CT code '386906001' = 'Capecitabine', given as 'Capecitabine (substance)'})</span></td><td>500 mg<span> (Details: UCUM code mg = 'mg')</span>/1 TAB<span> (Details: http://hl7.org/fhir/v3/orderableDrugForm code TAB = 'Tablet')</span></td></tr></table><blockquote><p><b>package</b></p><p><b>container</b>: Bottle - unit of produce usage (qualifier value) <span>(Details : {SNOMED CT code '419672006' = 'Bottle', given as 'Bottle - unit of produce usage (qualifier value)'})</span></p><h3>Contents</h3><table><tr><td>-</td><td><b>Item[x]</b></td></tr><tr><td>*</td><td>Capecitabine 500mg tablets (product) <span>(Details : {SNOMED CT code '134622004' = 'Capecitabine 500mg tablet', given as 'Capecitabine 500mg tablets (product)'})</span></td></tr></table><h3>Batches</h3><table><tr><td>-</td><td><b>LotNumber</b></td><td><b>ExpirationDate</b></td></tr><tr><td>*</td><td>9494788</td><td>22/05/2017</td></tr></table></blockquote></div>"
+  },
+  "code": {
+    "coding": [
+      {
+        "system": "http://www.nlm.nih.gov/research/umls/rxnorm",
+        "code": "213293",
+        "display": "Capecitabine 500mg oral tablet (Xeloda)"
+      }
+    ]
+  }
+}
