+++ lib/modules/argonaut/test/fixtures/observation.json
@@ -1,0 +1,46 @@
+{
+  "id": "a421b98b-5fd7-4595-9a12-486a1aa41e67",
+  "meta": {
+    "profile": [
+      "http://fhir.org/guides/argonaut/StructureDefinition/argo-observationresults"
+    ]
+  },
+  "status": "final",
+  "category": {
+    "coding": [
+      {
+        "system": "http://hl7.org/fhir/observation-category",
+        "code": "vital-signs",
+        "display": "vital-signs"
+      },
+      {
+        "system": "http://hl7.org/fhir/observation-category",
+        "code": "laboratory",
+        "display": "vital-signs"
+      }
+    ],
+    "text": "vital-signs"
+  },
+  "code": {
+    "coding": [
+      {
+        "system": "http://loinc.org",
+        "code": "72166-2",
+        "display": "Smoking Status"
+      }
+    ],
+    "text": "Smoking Status"
+  },
+  "subject": {
+    "reference": "Patient/example"
+  },
+  "effectiveDateTime": "2010-03-01",
+  "issued": "2010-03-01T15:30:10+01:00",
+  "valueQuantity": {
+    "value": 150.43678997612926,
+    "unit": "cm",
+    "system": "http://unitsofmeasure.org",
+    "code": "cm"
+  },
+  "resourceType": "Observation"
+}
