+++ lib/modules/argonaut/test/fixtures/patient.json
@@ -1,0 +1,186 @@
+{
+  "id": "7ec866ca-4130-4a40-82a9-b67f2626a5df",
+  "text": {
+    "status": "generated",
+    "div": "<div xmlns=\"http://www.w3.org/1999/xhtml\">Generated by <a href=\"https://github.com/synthetichealth/synthea\">Synthea</a>.Version identifier: v1.3.1-462-ga4e4ad9\n .   Person seed: -6604010938320047939  Population seed: 1521224923924</div>"
+  },
+  "extension": [
+    {
+      "url": "http://hl7.org/fhir/StructureDefinition/us-core-race",
+      "valueCodeableConcept": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/v3/Race",
+            "code": "2106-3",
+            "display": "White"
+          }
+        ],
+        "text": "White"
+      }
+    },
+    {
+      "url": "http://hl7.org/fhir/StructureDefinition/us-core-ethnicity",
+      "valueCodeableConcept": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/v3/Ethnicity",
+            "code": "2186-5",
+            "display": "Not Hispanic or Latino"
+          }
+        ],
+        "text": "Not Hispanic or Latino"
+      }
+    },
+    {
+      "url": "http://hl7.org/fhir/StructureDefinition/patient-mothersMaidenName",
+      "valueString": "Kimi714 Hermiston71"
+    },
+    {
+      "url": "http://hl7.org/fhir/us/core/StructureDefinition/us-core-birthsex",
+      "valueCode": "F"
+    },
+    {
+      "url": "http://hl7.org/fhir/StructureDefinition/birthPlace",
+      "valueAddress": {
+        "city": "Mansfield",
+        "state": "Massachusetts",
+        "country": "US"
+      }
+    },
+    {
+      "url": "http://synthetichealth.github.io/synthea/disability-adjusted-life-years",
+      "valueDecimal": 1.9189183066996247
+    },
+    {
+      "url": "http://synthetichealth.github.io/synthea/quality-adjusted-life-years",
+      "valueDecimal": 51.081081693300376
+    }
+  ],
+  "identifier": [
+    {
+      "system": "https://github.com/synthetichealth/synthea",
+      "value": "10a389cd-3422-4a98-9481-7eceec6118a9"
+    },
+    {
+      "type": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/v2/0203",
+            "code": "MR"
+          }
+        ]
+      },
+      "system": "http://hospital.smarthealthit.org",
+      "value": "10a389cd-3422-4a98-9481-7eceec6118a9"
+    },
+    {
+      "type": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/identifier-type",
+            "code": "SB"
+          }
+        ]
+      },
+      "system": "http://hl7.org/fhir/sid/us-ssn",
+      "value": "999-40-6899"
+    },
+    {
+      "type": {
+        "coding": [
+          {
+            "system": "http://hl7.org/fhir/v2/0203",
+            "code": "DL"
+          }
+        ]
+      },
+      "system": "urn:oid:2.16.840.1.113883.4.3.25",
+      "value": "S99980352"
+    }
+  ],
+  "name": [
+    {
+      "use": "official",
+      "family": [
+        "Ernser583"
+      ],
+      "given": [
+        "Rhea34"
+      ],
+      "prefix": [
+        "Mrs."
+      ]
+    },
+    {
+      "use": "maiden",
+      "family": [
+        "Marks830"
+      ],
+      "given": [
+        "Rhea34"
+      ],
+      "prefix": [
+        "Mrs."
+      ]
+    }
+  ],
+  "telecom": [
+    {
+      "system": "phone",
+      "value": "555-317-4311",
+      "use": "home"
+    }
+  ],
+  "gender": "female",
+  "birthDate": "1964-02-10",
+  "address": [
+    {
+      "extension": [
+        {
+          "extension": [
+            {
+              "url": "latitude",
+              "valueDecimal": -71.281889
+            },
+            {
+              "url": "longitude",
+              "valueDecimal": 42.499295000000004
+            }
+          ],
+          "url": "http://hl7.org/fhir/StructureDefinition/geolocation"
+        }
+      ],
+      "line": [
+        "554 Cummings Lane Suite 22"
+      ],
+      "city": "Bedford",
+      "state": "Massachusetts",
+      "postalCode": "01730",
+      "country": "US"
+    }
+  ],
+  "maritalStatus": {
+    "coding": [
+      {
+        "system": "http://hl7.org/fhir/v3/MaritalStatus",
+        "code": "M"
+      }
+    ]
+  },
+  "multipleBirthBoolean": false,
+  "communication": [
+    {
+      "language": {
+        "coding": [
+          {
+            "system": "urn:ietf:bcp:47",
+            "code": "en-US",
+            "display": "English"
+          }
+        ],
+        "text": "English"
+      }
+    }
+  ],
+  "resourceType": "Patient"
+}
