+++ lib/modules/argonaut/test/fixtures/procedure.json
@@ -1,0 +1,22 @@
+{
+    "id": "92134275-3ce9-4867-a2e9-49d712f91a1e",
+    "subject": {
+      "reference": "Patient/example"
+    },
+    "status": "completed",
+    "code": {
+      "coding": [
+        {
+          "system": "http://snomed.info/sct",
+          "code": "755621000000101",
+          "display": "Replacement of subcutaneous contraceptive"
+        }
+      ],
+      "text": "Replacement of subcutaneous contraceptive"
+    },
+    "performedPeriod": {
+      "start": "2010-08-09",
+      "end": "2010-08-09"
+    },
+    "resourceType": "Procedure"
+  }
